functions/measurement/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
arrow==1.3.0
    # via -r requirements.in
dnspython==2.4.2
    # via pymongo
pymongo==4.6.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via arrow
python-dotenv==1.0.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
types-python-dateutil==2.8.19.14
    # via arrow
